Add vh-check library
fixes title not displaying in view on mobile safari

for certain browsers (looking at you mobile safari) vh doesnt
consider the browser chrome in its calculations.

this change enables a css variable called --vh-offset that is
set when needed. the min-height of the body is now set to
calc(100vh - var(--vh-offset))